fix: add status to Payment fillable, cast discount amount to int
Two bugs found during end-to-end payment testing with discount codes:

1. Payment.status was missing from $fillable — confirmPayment() silently
   failed to update status from 'pending' to 'completed'. Added 'status'
   to the fillable array.

2. DiscountCodeService::apply() requires int $originalAmountPence but
   Payment.amount is cast to decimal:2 (float). Adding discount_amount
   (int) to amount (float) produced a float, causing a TypeError. Cast
   to (int) at the call site in confirmPayment().

+++ app/Http/Controllers/Api/PaymentController.php
@@ -354,7 +354,7 @@ class PaymentController extends Controller
                             $payment->discountCode,
                             $user->id,
                             $payment->id,
-                            $payment->amount + $payment->discount_amount
+                            (int) ($payment->amount + $payment->discount_amount)
                         );
                     } catch (\Exception $e) {
                         Log::error('Failed to apply discount code usage', [
